third_party/WebKit/LayoutTests/http/tests/devtools/profiler/cpu-profiler-save-load-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that CPU profiling is able to save/load.

Profiler was enabled.

Running: testSave
PASS: open was called with extension '.cpuprofile'
PASS: write was called with data length more than zero
PASS: close was called

Running: testLoad
PASS: the content of loaded profile matches with the original profile.

Profiler was disabled.
